KB_T03_008-add lat&lon to enum city

+++ src/main/java/com/kb/windsurfersweatherservice/model/Weather.java
@@ -12,8 +12,8 @@ import lombok.ToString;
 public class Weather {
 
     private String cityName;
-    private float temperature;
-    private float windSpeed;
-    private float lat;
-    private float lon;
+    private double temperature;
+    private double windSpeed;
+    private double lat;
+    private double lon;
 }
